Move camera logic to separate service, add shutdown options if exception will throw
Use this links to repair a problem with sending broadcast: https://stackoverflow.com/questions/76919130/android-14-context-registered-broadcast-receivers-not-working

https://issuetracker.google.com/issues/293487554?pli=1
