+++ app/src/main/java/com/example/safedistance/utils/Constants.kt
@@ -1,0 +1,11 @@
+package com.example.safedistance.utils
+
+enum class Constants(name: String) {
+    ACTION_CLOSE_ALL_ACTIVITIES("CLOSE_ALL_ACTIVITIES"),
+    ACTION_CLOSE_ALL_SERVICES("CLOSE_ALL_SERVICES"),
+    ACTION_DISTANCE("DISTANCE"),
+
+    VALUE_DISTANCE("distance"),
+    VALUE_TITLE("title"),
+    VALUE_MESSAGE("message")
+}
